fix: add deps for core format
